+++ lib/pages/book/merchants_page.dart
@@ -5,6 +5,8 @@ import '../../manager/app_config_manager.dart';
 import '../../manager/l10n_manager.dart';
 import '../../models/vo/user_book_vo.dart';
 import '../../widgets/common/common_simple_crud_list.dart';
+import '../../models/dto/item_filter_dto.dart';
+import '../../routes/app_routes.dart';
 
 class MerchantsPage extends StatelessWidget {
   const MerchantsPage({super.key, required this.accountBook});
@@ -31,6 +33,13 @@ class MerchantsPage extends StatelessWidget {
           accountBook.id,
           item.id,
         ),
+        onItemTap: (item) {
+          final filter = ItemFilterDTO(shopCodes: [item.code]);
+          Navigator.of(context).pushNamed(
+            AppRoutes.items,
+            arguments: [accountBook, filter, item.name],
+          );
+        },
       ),
     );
   }